@@ -39,6 +39,7 @@
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-0": "^6.5.0",
     "babel-register": "^6.9.0",
+    "bcrypt": "^0.8.7",
     "classlist-polyfill": "^1.0.3",
     "css-loader": "^0.23.1",
     "css-modules-require-hook": "^4.0.1",
@@ -52,6 +53,7 @@
     "good-console": "^6.1.2",
     "h2o2": "^5.1.0",
     "hapi": "^13.5.0",
+    "hapi-auth-basic": "^4.2.0",
     "hapi-require-https": "^2.0.9",
     "inert": "^4.0.1",
     "inject-then": "^2.0.5",
@@ -93,7 +95,7 @@
     "chai-things": "^0.2.0",
     "concurrently": "^2.2.0",
     "enzyme": "^2.4.1",
-    "eslint": "^3.7.1",
+    "eslint": "^3.8.1",
     "lint-staged": "^2.0.2",
     "mocha": "^2.5.3",
     "mocha-webpack": "^0.4.0",
