@@ -1,8 +1,9 @@
 import React from 'react';
 import styles from './heading2.scss';
+import getClassnames from 'helpers/getClassNames';
 
-const Heading2 = ({ children, text }) => (
-  <h2 className={styles.root}>
+const Heading2 = ({ children, text, size }) => (
+  <h2 className={getClassnames(styles, ['root', size])}>
     {children || text}
   </h2>
 );
