@@ -1,15 +1,19 @@
 import React from 'react';
-import Heading1 from 'components/Heading1';
+import Container from 'components/Container';
 
 import styles from './hero.scss';
 
 const Hero = ({ title, missionTitle, missionDescription, background }) => (
   <section>
     <div className={styles.top}>
-      <div
-        style={{ backgroundImage: `url(${background})` }}
-        className={styles.background} />
-      <Heading1 text={title} />
+      <Container>
+        <div className={styles.inner}>
+          <div
+            style={{ backgroundImage: `url(${background})` }}
+            className={styles.background} />
+          <h1>{title}</h1>
+        </div>
+      </Container>
     </div>
   </section>
 );
