@@ -1,10 +1,11 @@
 import React from 'react';
 import Markdown from 'react-markdown';
 import styles from 'components/Paragraph/paragraph.scss';
+import getClassNames from 'helpers/getClassNames';
 
-const MarkdownParagraph = ({ text, noMargin }) => (
+const MarkdownParagraph = ({ text, margin }) => (
   <Markdown
-    className={styles[noMargin ? 'no-margin' : 'root']}
+    className={getClassNames(styles, ['root', margin])}
     disallowedTypes={['HtmlBlock', 'Image']}
     source={text} />
 );
