@@ -1,8 +1,9 @@
 import React from 'react';
 import styles from './paragraph.scss';
+import getClassNames from 'helpers/getClassNames';
 
-const Paragraph = ({ children, text, noMargin }) => (
-  <p className={styles[noMargin ? 'no-margin' : 'root']}>
+const Paragraph = ({ children, text, margin, size }) => (
+  <p className={getClassNames(styles, ['root', margin, size])}>
     {children || text}
   </p>
 );
