@@ -3,14 +3,14 @@ import styles from './preview-petition.scss';
 import PetitionActions from 'components/PetitionActions';
 import Petition from 'components/Petition';
 
-const PreviewPetition = ({ petition, fetchPetition, publishPetition }) => (
+const PreviewPetition = ({ petition, publishPetition }) => (
   <div className={styles.root}>
-    <Petition preview />
+    <Petition petition={petition} preview />
     {!petition.published &&
       <div className={styles.actions}>
         <PetitionActions
           petition={petition}
-          actions={{fetchPetition, publishPetition}}
+          actions={{publishPetition}}
         />
       </div>
     }
