@@ -1,18 +1,29 @@
 import React from 'react';
 import settings from 'settings';
 import FormLayout from 'components/FormLayout';
-import TrustConfirmationForm from 'components/TrustConfirmationForm';
+import TrustSupportConfirmationForm from 'components/TrustSupportConfirmationForm';
+import TrustPublishConfirmationForm from 'components/TrustPublishConfirmationForm';
 
 const TrustConfirmation = ({
   isLoggedIn,
-  me
+  me,
+  action
 }) => (
   <FormLayout
     title={settings.trustConfirmationPage.title}
     intro={settings.trustConfirmationPage.intro}>
     <p>Thanks {me.firstname}, we sent an SMS verification code to <b>{me.mobile}</b></p>
-    <TrustConfirmationForm />
+    {action === 'support' &&
+      <TrustSupportConfirmationForm />
+    }
+    {action === 'publish' &&
+      <TrustPublishConfirmationForm />
+    }
   </FormLayout>
 );
+
+TrustConfirmation.propTypes = {
+  action: React.PropTypes.oneOf(['support', 'publish'])
+};
 
 export default TrustConfirmation;
