@@ -1,14 +1,14 @@
 import React from 'react';
 import { reduxForm } from 'redux-form';
+import { supportPetition, resendVerification } from 'actions/SupportActions';
 import trustConfirmationValidator from 'form/trustConfirmationValidator';
-import assignUserData from 'form/assignUserData';
+import assignUserAndPetitionData from 'form/assignUserAndPetitionData';
 import Fieldset from 'components/Fieldset';
 import FormFieldsIterator from 'components/FormFieldsIterator';
 import Button from 'components/Button';
 import ButtonLink from 'components/ButtonLink';
 import ButtonSet from 'components/ButtonSet';
 import FIELDS from './fields';
-import { supportPetition, resendVerification } from 'actions/SupportActions';
 import trustForm from 'selectors/trustForm';
 import hasValidUserData from 'helpers/hasValidUserData';
 
@@ -17,11 +17,11 @@ const TrustConfirmationForm = ({
   handleSubmit,
   resendVerification,
   me,
-  petitionId,
+  petition,
   submitting
 }) => (
   <form onSubmit={handleSubmit((values, dispatch) => supportPetition(
-    assignUserData(values, me), dispatch)
+    assignUserAndPetitionData(values, me, petition), dispatch)
   )}>
     <Fieldset>
       <FormFieldsIterator
@@ -32,19 +32,19 @@ const TrustConfirmationForm = ({
     <Fieldset modifier={'actions'}>
       <ButtonSet>
         <ButtonLink
-          href={`/trust/support/${petitionId}`}
+          href={`/trust/support/${petition.id}`}
           text={'Back to details'}
         />
         <Button
-          onClick={() => resendVerification(petitionId, me)}
+          onClick={() => resendVerification(petition, me)}
           disabled={!hasValidUserData(me)}
           type={'button'}
-          text={'Se-rend SMS'}
+          text={'Re-send SMS'}
         />
         <Button
           text={'Complete verification'}
           modifier={'accent'}
-          disabled={submitting || !fields._meta.allValid}
+          disabled={!hasValidUserData(me) || submitting || !fields._meta.allValid}
         />
       </ButtonSet>
     </Fieldset>
@@ -55,15 +55,15 @@ TrustConfirmationForm.propTypes = {
   fields: React.PropTypes.object.isRequired,
   handleSubmit: React.PropTypes.func.isRequired,
   resendVerification: React.PropTypes.func.isRequired,
+  petition: React.PropTypes.object.isRequired,
   me: React.PropTypes.object.isRequired,
-  petitionId: React.PropTypes.string.isRequired,
   submitting: React.PropTypes.bool.isRequired
 };
 
 export const mapStateToProps = ({ petition, me, trust }) => trustForm(petition, me, trust);
 
 export const mapDispatchToProps = (dispatch) => ({
-  resendVerification: (petitionId, user) => dispatch(resendVerification({ petitionId, user }))
+  resendVerification: (petition, user) => dispatch(resendVerification({ petition, user }))
 });
 
 export default reduxForm({
