@@ -1,27 +1,35 @@
 import React from 'react';
 import { reduxForm } from 'redux-form';
 import { supportPetition } from 'actions/SupportActions';
-import trustValidator from 'form/trustValidator';
+import trustSupportValidator from 'form/trustSupportValidator';
+import assignPetitionData from 'form/assignPetitionData';
 import Fieldset from 'components/Fieldset';
 import FormFieldsIterator from 'components/FormFieldsIterator';
 import Button from 'components/Button';
 import ButtonLink from 'components/ButtonLink';
 import ButtonSet from 'components/ButtonSet';
-import FIELDS_SUPPORTING from './fieldsForSupporting';
+import FIELDS from './fields';
 import trustForm from 'selectors/trustForm';
 
-const TrustForm = ({ fields, handleSubmit, submitting, petitionId }) => (
-  <form onSubmit={handleSubmit(supportPetition)}>
+const TrustSupportForm = ({
+  fields,
+  handleSubmit,
+  petition,
+  submitting
+}) => (
+  <form onSubmit={handleSubmit((values, dispatch) => supportPetition(
+    assignPetitionData(values, petition), dispatch)
+  )}>
     <Fieldset>
       <FormFieldsIterator
         reduxFormFields={fields}
-        fieldsArray={FIELDS_SUPPORTING}
+        fieldsArray={FIELDS}
       />
     </Fieldset>
     <Fieldset modifier={'actions'}>
       <ButtonSet>
         <ButtonLink
-          href={`/petitions/${petitionId}`}
+          href={`/petitions/${petition.id}`}
           text={'Back to petition'}
         />
         <Button
@@ -34,16 +42,18 @@ const TrustForm = ({ fields, handleSubmit, submitting, petitionId }) => (
   </form>
 );
 
-TrustForm.propTypes = {
+TrustSupportForm.propTypes = {
   fields: React.PropTypes.object.isRequired,
   handleSubmit: React.PropTypes.func.isRequired,
+  petition: React.PropTypes.object.isRequired,
+  me: React.PropTypes.object.isRequired,
   submitting: React.PropTypes.bool.isRequired
 };
 
 export const mapStateToProps = ({ petition, me, trust }) => trustForm(petition, me, trust);
 
 export default reduxForm({
-  form: 'trust',
-  fields: FIELDS_SUPPORTING.map(field => field.name),
-  validate: trustValidator
-}, mapStateToProps)(TrustForm);
+  form: 'trustSupport',
+  fields: FIELDS.map(field => field.name),
+  validate: trustSupportValidator
+}, mapStateToProps)(TrustSupportForm);
