@@ -7,12 +7,15 @@ import { showFlashMessage } from 'actions/FlashActions';
 import settings from 'settings';
 import TrustConfirmation from 'components/TrustConfirmation';
 import getPetitionPath from 'selectors/petitionPath';
+import getPetitionForm from 'selectors/petitionForm';
 import trustSubmitted from 'helpers/trustSubmitted';
 import hasValidUserData from 'helpers/hasValidUserData';
 
-const TrustConfirmationContainer = withRouter(React.createClass({
+const TrustSupportConfirmationContainer = withRouter(React.createClass({
   componentWillMount () {
     // const { router, petition, hasValidUserData, showFlashMessage } = this.props;
+
+    // If there is no valid user data, redirect to support form
     // if (!hasValidUserData) {
     //   showFlashMessage(settings.flashMessages.invalidUserDataError, 'error');
     //   router.push(`/trust/support/${petition.id}`);
@@ -21,6 +24,7 @@ const TrustConfirmationContainer = withRouter(React.createClass({
 
   componentWillUpdate (nextProps) {
     const { router, petition, trustSubmitted, isTrustedUser } = nextProps;
+
     // If we have submitted trust successfully
     if (trustSubmitted && isTrustedUser) {
       router.push(getPetitionPath(petition));
@@ -31,18 +35,19 @@ const TrustConfirmationContainer = withRouter(React.createClass({
     return (
       <div>
         <Helmet title={settings.trustConfirmationPage.title} />
-        <TrustConfirmation {...this.props} />
+        <TrustConfirmation {...this.props} action={'support'} />
       </div>
     );
   }
 }));
 
-TrustConfirmationContainer.fetchData = ({ store, params }) => {
+TrustSupportConfirmationContainer.fetchData = ({ store, params }) => {
   return store.dispatch(fetchPetition(params.id));
 };
 
 export const mapStateToProps = ({ petition, trust, me }) => ({
   me,
+  petition: getPetitionForm(petition),
   trustSubmitted: trustSubmitted(petition, trust),
   isTrustedUser: trust.isTrustedUser,
   hasValidUserData: hasValidUserData(me)
@@ -52,8 +57,9 @@ export const mapDispatchToProps = (dispatch) => ({
   showFlashMessage: (message, type) => dispatch(showFlashMessage(message, type))
 });
 
-TrustConfirmationContainer.propTypes = {
+TrustSupportConfirmationContainer.propTypes = {
   me: React.PropTypes.object.isRequired,
+  petition: React.PropTypes.object.isRequired,
   trustSubmitted: React.PropTypes.bool.isRequired,
   isTrustedUser: React.PropTypes.bool.isRequired,
   hasValidUserData: React.PropTypes.bool.isRequired
@@ -62,4 +68,4 @@ TrustConfirmationContainer.propTypes = {
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(TrustConfirmationContainer);
+)(TrustSupportConfirmationContainer);
