@@ -6,13 +6,26 @@ import { createStore, applyMiddleware } from 'redux';
 import thunkMiddleware from 'redux-thunk';
 import { Provider } from 'react-redux';
 import reducers from 'reducers';
-
 import routes from 'routes';
+
 import stringifyHeadData from 'server/stringifyHeadData';
 import getBundles from 'server/getBundles';
 
-// Needed to prefix pageview events for analytics.
-const SITE_NAME = process.env.SITE_NAME;
+/*
+ * Note: this file contains server-side rendering logic, called from server.js.
+ *
+ * YOU SHOULD NOT NEED TO EDIT THIS FILE unless adding additional server logic /
+ * template variables. Data fetching can be handled in route handler containers.
+ *
+ * Get data on the server by adding a `fetchData` function to a container e.g.
+ *
+ * PetitionContainer.fetchData = ({ store, params }) => {
+ *   return store.dispatch(fetchPetition(params.id));
+ * };
+ *
+ * This function is run on the server before rendering, dispatching an action
+ * to pre-fill the Redux store with data for that route.
+ */
 
 export default (request, reply, next) => {
   match({ routes: routes(), location: { pathname: request.path, query: request.query } }, (error, redirectLocation, renderProps) => {
@@ -23,6 +36,7 @@ export default (request, reply, next) => {
     } else if (renderProps == null) {
       reply('Not found').code(404);
     } else {
+      // Set initial state
       const initialState = {};
 
       // Create our store
@@ -34,16 +48,17 @@ export default (request, reply, next) => {
 
       // Get the component tree
       const components = renderProps.components || [];
-      // Extract our page component
+      // Extract our page component for this route
       const Component = components[components.length - 1];
-      // Extract `fetchData` if exists
+      // Extract `fetchData` from component if exists, otherwise return empty promise
       const fetchData = (Component && Component.fetchData) || (() => Promise.resolve());
       // Get from renderProps
       const { location, params, history } = renderProps;
 
-      // Fetch async data, then render
+      // Run fetchData to get async data, then render
       fetchData({ store, location, params, history })
         .then(() => {
+          // Construct our markup
           const reactString = ReactDOMServer.renderToString(
             <Provider store={store}>
               <RouterContext {...renderProps} />
@@ -61,7 +76,7 @@ export default (request, reply, next) => {
             initialState: JSON.stringify(initialState),
             head: stringifyHeadData(headData),
             bundles: getBundles(),
-            siteName: SITE_NAME
+            siteName: process.env.SITE_NAME // analytics & tracking
           }));
         })
         .catch((err) => {
