@@ -5,7 +5,7 @@ import path from 'path';
 export default {
   find: (id) => {
     const requestPath = path.join('/petitions', id.toString());
-    const requestParams = { resolve: 'city,owner,links', extend: 'supporting' };
+    const requestParams = { resolve: 'city,owner,links,mentions', extend: 'supporting' };
     return ApiClient.request(requestPath, requestParams);
   },
 
@@ -37,24 +37,20 @@ export default {
     return ApiClient.request(requestPath, { data: petition }, 'POST');
   },
 
-  publish: (petition) => {
+  publish: ({ petition, mobile_token }) => {
     const requestPath = path.join('/petitions', petition.id.toString(), '/event/publish');
-    return ApiClient.request(requestPath, null, POST);
+    const payload = mobile_token ? { mobile_token } : {}; // eslint-disable-line camelcase
+    return ApiClient.request(requestPath, { data: payload }, POST);
   },
 
-  support: (trustData) => {
-    const requestPath = path.join('/petitions', trustData.petitionId.toString(), '/event/support');
-    return ApiClient.request(requestPath, { data: trustData }, POST);
+  support: ({ petition, mobile_token, user }) => {
+    const requestPath = path.join('/petitions', petition.id.toString(), '/event/support');
+    const payload = mobile_token ? { mobile_token, user } : { user }; // eslint-disable-line camelcase
+    return ApiClient.request(requestPath, { data: payload }, POST);
   },
 
-  respond: (response) => {
-    const requestPath = path.join('/petitions', response.petitionId.toString(), 'event/setFeedback?resolve=city');
-    const requestPayload = {
-      data: {
-        answer: response.answer,
-        token: response.token
-      }
-    };
-    return ApiClient.request(requestPath, requestPayload, POST);
+  respond: ({ petitionId, answer, token }) => {
+    const requestPath = path.join('/petitions', petitionId.toString(), 'event/setFeedback?resolve=city');
+    return ApiClient.request(requestPath, { data: { answer, token } }, POST);
   }
 };
